+++ bot/mod_commands/setgame.js
@@ -1,0 +1,9 @@
+'use strict';
+
+module.exports = {
+	desc: "Set game",
+	usage: "<game name>",
+	process: (bot, msg, suffix, flags) => {
+		bot.user.setGame(suffix);
+	}
+}